Printing positions to client
