+++ src/perception/__init__.py
@@ -1,5 +1,6 @@
 from .sensing import Sensing
 from .state import State
 from .map import Map
+from .clientMap import ClientMap
 
 from .mapTest import TestMap
