+++ src/robot/robot.py
@@ -85,6 +85,7 @@ class Robot:
 
         right_encoder, left_encoder, obstacle_distance = self.sensing.collect()
         self.state.update_from_sensors(right_encoder, left_encoder, obstacle_distance)
+        self.network.update_position_message(self.state)
         command = self.decision_making.decide(self.state, self.target, self.target_node)
         command.execute(self.state)
         if self.debug:
